fix some compile time warnings

+++ src/serial/serialcommands.cpp
@@ -177,7 +177,7 @@ void printState() {
 		wifiNetwork.getAddress().toString().c_str(),
 		WiFi.macAddress().c_str(),
 		statusManager.getStatus(),
-		wifiNetwork.getWiFiState()
+		static_cast<int>(wifiNetwork.getWiFiState())
 	);
 
 	logger.info("%s", FULL_VENDOR_STR);
@@ -308,7 +308,7 @@ void cmdGet(CmdParser* parser) {
 			wifiNetwork.getAddress().toString().c_str(),
 			WiFi.macAddress().c_str(),
 			statusManager.getStatus(),
-			wifiNetwork.getWiFiState()
+			static_cast<int>(wifiNetwork.getWiFiState())
 		);
 		auto& sensor0 = sensorManager.getSensors()[0];
 		sensor0->motionLoop();
